Move legacy testing system into legacy-2-x subfolder.
This is in prep for the work needed to support the new testing API's
proposed in emberjs/rfcs#232.
