+++ addon-test-support/ember-qunit/legacy-2-x/module-for-component.js
